[TASK] Combined existing files and queue in MultiUploadViewHelper
This change currently uses jQuery to manipulate the DOM in a way that 
places the editable file list part inside the queue of the plupload instance. 
Look and sequence of all file rows are preserved after files are moved from 
the queue to the existing file list.

Dramatically decreases the page real estate MultiUploadViewHelper takes.