[TASK] Create a branch for "unbuffered queries"
This feature was introduced because of performance problems with the
AGES PGE. Instead of introducing such a complicated feature rather
the mysql indexes should have get optimized.

Now as this feature isn't senseless at all create a branch for it
which holds the modifications.

But instead of hacking the feature into the existing code the whole
concept should get changed. NEW CONCEPT:
Create an event queue and a "select" method/object which allows to
retrieve tasks which shall get processed. Start the whole stuff by
putting the "main" query into the queue. When results are ready
sub-queries and rendering-requests will get put into the queue until
everything is finished.