[BUGFIX] Extension breaks on usage of "Show content from Page"
When TYPO3 features like "Show content from page" or the
"Insert records" content element have been use which show
a container residing on a different page than the currently
rendered one a fatal error was generated in the frontend.

The reason for this is the kb_nescefe ContentRepository. It only
took the currently rendered page into account for retrieving
content elements.

This patch fixes the issue by taking the TSFE->contentPid into
account and handling "Insert records" content elements as good
as possible

Resolves: #66521
Releases: master